requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
flask==3.0.2